# evidencias/M8_openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Asistente documental RAG",
        "description": "Responde preguntas solo con los documentos internos, citando las fuentes.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/documents": {
            "get": {
                "tags": [
                    "documentos"
                ],
                "summary": "List Documents",
                "description": "Fuentes indexadas con su nº de chunks y su origen (corpus o subida).",
                "operationId": "list_documents_documents_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/DocumentOut"
                                    },
                                    "type": "array",
                                    "title": "Response List Documents Documents Get"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "documentos"
                ],
                "summary": "Upload Documents",
                "description": "Sube 1..n archivos a data/uploads/ y los ingiere (una subida previa con el mismo nombre\nse reemplaza; un nombre del corpus devuelve 409).",
                "operationId": "upload_documents_documents_post",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_upload_documents_documents_post"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IngestReportOut"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorOut"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorOut"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorOut"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorOut"
                                }
                            }
                        }
                    },
                    "413": {
                        "description": "Request Entity Too Large",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorOut"
                                }
                            }
                        }
                    },
                    "415": {
                        "description": "Unsupported Media Type",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorOut"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorOut"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorOut"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorOut"
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorOut"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorOut"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/ingest": {
            "post": {
                "tags": [
                    "documentos"
                ],
                "summary": "Reingest",
                "description": "Re-ingiere data/docs + data/uploads (con {\"reset\": true} vacía antes el índice).",
                "operationId": "reingest_ingest_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "anyOf": [
                                    {
                                        "$ref": "#/components/schemas/IngestRequest"
                                    },
                                    {
                                        "type": "null"
                                    }
                                ],
                                "title": "Body"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IngestReportOut"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorOut"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorOut"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorOut"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorOut"
                                }
                            }
                        }
                    },
                    "413": {
                        "description": "Request Entity Too Large",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorOut"
                                }
                            }
                        }
                    },
                    "415": {
                        "description": "Unsupported Media Type",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorOut"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorOut"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorOut"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorOut"
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorOut"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorOut"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Body_upload_documents_documents_post": {
                "properties": {
                    "files": {
                        "items": {
                            "type": "string",
                            "contentMediaType": "application/octet-stream"
                        },
                        "type": "array",
                        "title": "Files",
                        "description": "Archivos .txt, .md o .pdf (≤ 10 MB cada uno)"
                    }
                },
                "type": "object",
                "required": [
                    "files"
                ],
                "title": "Body_upload_documents_documents_post"
            },
...
            "DocumentOut": {
                "properties": {
                    "source": {
                        "type": "string",
                        "title": "Source"
                    },
                    "chunks": {
                        "type": "integer",
                        "title": "Chunks"
                    },
                    "origin": {
                        "type": "string",
                        "enum": [
                            "corpus",
                            "upload",
                            "desconocido"
                        ],
                        "title": "Origin"
                    }
                },
                "type": "object",
                "required": [
                    "source",
                    "chunks",
                    "origin"
                ],
                "title": "DocumentOut",
                "description": "Fuente indexada con su nº de chunks y su origen."
            },
            "ErrorOut": {
                "properties": {
                    "error": {
                        "type": "string",
                        "title": "Error"
                    },
                    "detail": {
                        "type": "string",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "required": [
                    "error",
                    "detail"
                ],
                "title": "ErrorOut",
                "description": "Formato único de error."
            },
...
            "IngestReportOut": {
                "properties": {
                    "files_processed": {
                        "type": "integer",
                        "title": "Files Processed"
                    },
                    "files_skipped": {
                        "type": "integer",
                        "title": "Files Skipped"
                    },
                    "documents": {
                        "type": "integer",
                        "title": "Documents"
                    },
                    "chunks_added": {
                        "type": "integer",
                        "title": "Chunks Added"
                    },
                    "total_in_store": {
                        "type": "integer",
                        "title": "Total In Store"
                    },
                    "sources": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Sources"
                    },
                    "skipped": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Skipped"
                    },
                    "duration_s": {
                        "type": "number",
                        "title": "Duration S"
                    }
                },
                "type": "object",
                "required": [
                    "files_processed",
                    "files_skipped",
                    "documents",
                    "chunks_added",
                    "total_in_store",
                    "sources",
                    "skipped",
                    "duration_s"
                ],
                "title": "IngestReportOut",
                "description": "Reporte de ingesta."
            },
            "IngestRequest": {
                "properties": {
                    "reset": {
                        "type": "boolean",
                        "title": "Reset",
                        "default": false
                    }
                },
                "type": "object",
                "title": "IngestRequest",
                "description": "Opciones de re-ingesta."
            },
...
        }
    }
}